refactor: add table package
