+++ src/main/java/kitchenpos/table/domain/OrderTables.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.table.domain;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import javax.persistence.Embeddable;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
-import kitchenpos.dto.OrderTableResponse;
+import kitchenpos.table.dto.OrderTableResponse;
 
 @Embeddable
 public class OrderTables {
